+++ postgres_v8.4.x/pg_mana/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2026-09-22
+### Added
+- `kill_idle_txn` function.
+- `kill_idle_sess` function.
+
+### Fixed
+- Using "quote_ident" in `get_mvidx_stmt` function.
+
+### Removed
+- The `kill_idle` function.
+
 ## [0.5.2] - 2026-09-20
 ### Fixed
 - Using the "quote_ident" function in queries that refer to object names.